DrawableObject::Draw override fix
Drawable Objects can now be added to a window's PXLM via PixelManager::AddObject(obj), and be drawn on the screen.

DrawableObject has methods for changing location in each of the four cardinal directions... though I am not sure if this is the most elegant way of handling location changes. Perhaps the location property should be public?

Exception is thrown and unhandled when objects travel off-screen.