Tickers Update
GPW.cpp & GPW.h :: Updated movement mechanics. Movement only occurs when the timer is off. A direction variable has been added to DrawableObjects which allows the user to change the direction of travel while the timer is running.
KNOWN ISSUE-- when timer is off, and player strikes a movement key: movement occurs immediately, and timer begins. However, player does not release key *immediately*. As such, direction is chosen for the next movement. TLDR it is impossible to move a single tile-- two tiles are always traversed.

DrawableObject.h :: A complex<int> indicating direction of travel has been added. There are functions for moving in each of the cardinal directions, and general movement function which moves in the complex<int> direction of travel. Additionally, a Compass namespace with universally referenceable directions of travel has been added.
KNOWN ISSUE-- This is inelegant and needs tidying.

PixelWindow.h & PixelWindow.cpp :: Added a clock to all PixelWindows. Derived classes are currently responsible for updating the clock each frame.
KNOWN ISSUE-- This is unideal. The ProcessFrame() virtual function should automatically call clock.Tick(). A deeper understanding of abstract classes is needed to realize this.

Tickers.hpp :: Addition of a clock and timer classes.
Clock is meant solely to keep track of how much time has elapsed, both from the start of the program and between each call of its Tick() function.
The Timer class is meant to keep track of whether a certain interval, defined at Timer initialization, has passed between two points in the program.